# spring-test/spring-test/src/main/resources/specification/car-service.openapi.yml
openapi: 3.0.3

info:
  title: Car service API
  description: OpenAPI specification based on Car model
  version: 0.0.1

servers:
  - url: http://localhost:8080/api/v1

tags:
  - name: car

paths:
  "/cars":
    get:
      security:
        - BearerAuth: []
      tags:
        - car
      summary: "Get all cars"
      description: "Fetch all existing cars"
      operationId: getCars
      parameters:
        - in: query
          name: producer
          required: false
          schema:
            type: string
      responses:
        '200':
          $ref: "#/components/responses/CarList"
        '400':
          description: Invalid request
    post:
      security:
        - BearerAuth: []
      tags:
        - car
      summary: "Create car"
      operationId: createCar
      requestBody:
        $ref: "#/components/requestBodies/Car"
      responses:
        '201':
          $ref: "#/components/responses/Car"
        '400':
          description: Invalid request
        '422':
           description: Validation exception
  "/cars/{id}":
    get:
      security:
        - BearerAuth: []
      tags:
        - car
      summary: "Get specific car"
      description: "Get car by id"
      operationId: getCar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: "Car id"
      responses:
        '200':
          $ref: "#/components/responses/Car"
        '400':
          description: Invalid request
    put:
      security:
        - BearerAuth: []
      tags:
        - car
      summary: "Modify specific car"
      description: "Update car properties"
      operationId: modifyCar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: "Car id"
      requestBody:
        $ref: "#/components/requestBodies/Car"
      responses:
        '201':
          $ref: "#/components/responses/Car"
        '400':
          description: Invalid request
        '422':
          description: Validation exception
    patch:
      security:
        - BearerAuth: []
      tags:
        - car
      summary: "Modify specific car partially"
      description: "Update only car properties included in the body"
      operationId: modifyCarPartially
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: "Car id"
      requestBody:
        $ref: "#/components/requestBodies/Car"
      responses:
        '201':
          $ref: "#/components/responses/Car"
        '400':
          description: Invalid request
        '422':
          description: Validation exception
    delete:
      security:
        - BearerAuth: []
      tags:
        - car
      summary: "Delete specific car"
      description: "Delete car by id"
      operationId: deleteCar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: "Car id"
      responses:
        '200':
          description: "Successful response. Returns the ID of the deleted car."
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedId:
                    type: integer
                    format: int64
                    example: 1
        #          $ref: "#/components/responses/Car"

        '400':
          description: Invalid request

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Car:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 5
        model:
          type: string
          example: "GTI"
          minLength: 3
          maxLength: 64
        producer:
          type: string
          example: "Volkswagen"
          minLength: 3
          maxLength: 64
        power:
          type: integer
          format: int64
          minimum: 1
          maximum: 3000
          example: 250
  responses:
    Car:
      description: Car
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Car'
    CarList:
      description: Car list
      content:
        application/json:
          schema:
            type: object
            items:
              $ref: '#/components/schemas/Car'
  requestBodies:
    Car:
      description: Car to be created
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Car'